+++ mne/tests/test_evoked.py
@@ -1,5 +1,6 @@
 # Author: Alexandre Gramfort <gramfort@nmr.mgh.harvard.edu>
 #         Denis Engemann <d.engemann@fz-juelich.de>
+#         Andrew Dykstra <andrew.r.dykstra@gmail.com>
 #
 # License: BSD (3-clause)
 
@@ -14,8 +15,12 @@ from nose.tools import assert_true, assert_raises
 
 from mne import (equalize_channels, pick_types, read_evoked, write_evoked,
                  read_evokeds, write_evokeds)
-from mne.evoked import _get_peak
+from mne.evoked import _get_peak, EvokedArray
+from mne.epochs import EpochsArray
+
 from mne.utils import _TempDir, requires_pandas, requires_nitime
+
+from mne.io.meas_info import create_info
 
 warnings.simplefilter('always')
 
@@ -320,3 +325,46 @@ def test_equalize_channels():
     equalize_channels(my_comparison)
     for e in my_comparison:
         assert_equal(ch_names, e.ch_names)
+
+
+def test_array_epochs():
+    """Test creating evoked from array
+    """
+
+    # creating
+    rng = np.random.RandomState(42)
+    data1 = rng.randn(20, 60)
+    sfreq = 1e3
+    ch_names = ['EEG %03d' % (i + 1) for i in range(20)]
+    types = ['eeg'] * 20
+    info = create_info(ch_names, sfreq, types)
+    evoked1 = EvokedArray(data1, info, tmin=-0.01)
+
+    # save, read, and compare evokeds
+    tmp_fname = op.join(tempdir, 'evkdary-ave.fif')
+    evoked1.save(tmp_fname)
+    evoked2 = read_evokeds(tmp_fname)[0]
+    data2 = evoked2.data
+    assert_allclose(data1, data2)
+    assert_allclose(evoked1.times, evoked2.times)
+    assert_equal(evoked1.first, evoked2.first)
+    assert_equal(evoked1.last, evoked2.last)
+    assert_equal(evoked1.kind, evoked2.kind)
+    assert_equal(evoked1.nave, evoked2.nave)
+
+    # now compare with EpochsArray (with single epoch)
+    data3 = data1[np.newaxis, :, :]
+    events = np.c_[10, 0, 1]
+    evoked3 = EpochsArray(data3, info, events=events, tmin=-0.01).average()
+    assert_allclose(evoked1.data, evoked3.data)
+    assert_allclose(evoked1.times, evoked3.times)
+    assert_equal(evoked1.first, evoked3.first)
+    assert_equal(evoked1.last, evoked3.last)
+    assert_equal(evoked1.kind, evoked3.kind)
+    assert_equal(evoked1.nave, evoked3.nave)
+
+    # test match between channels info and data
+    ch_names = ['EEG %03d' % (i + 1) for i in range(19)]
+    types = ['eeg'] * 19
+    info = create_info(ch_names, sfreq, types)
+    assert_raises(ValueError, EvokedArray, data1, info, tmin=-0.01)